DBChat-CitizenConnect/requirements.txt
flask-cors
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
# langchain-text==0.2.2
langsmith==0.1.93
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.36.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==23.3.2
pydantic==2.8.2
pydantic_core==2.20.1
pyodbc==5.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tqdm==4.66.4
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4